👌 IMPROVE: Rename and move TodoApp class to lib directory
