+++ app/loader.php
@@ -3,4 +3,5 @@
 require_once 'config.php';
 require_once 'lib/helpers.php';
 require_once 'lib/Response.php';
-require_once 'app.php';
+
+require_once 'lib/TodoApp.php';
